+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/NoBlankLineBeforeRbraceRule.kt
@@ -22,7 +22,8 @@ class NoBlankLineBeforeRbraceRule : Rule("no-blank-line-before-rbrace") {
             if (split.size > 2) {
                 emit(
                     node.startOffset + split[0].length + split[1].length + 1,
-                    "Unexpected blank line(s) before \"}\"", true
+                    "Unexpected blank line(s) before \"}\"",
+                    true
                 )
                 if (autoCorrect) {
                     (node as LeafPsiElement).rawReplaceWithText("${split.first()}\n${split.last()}")