+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/ParameterListWrappingRule.kt
@@ -2,9 +2,13 @@ package com.pinterest.ktlint.ruleset.standard
 
 import com.pinterest.ktlint.core.KtLint
 import com.pinterest.ktlint.core.Rule
+import com.pinterest.ktlint.core.ast.ElementType.COLLECTION_LITERAL_EXPRESSION
 import com.pinterest.ktlint.core.ast.ElementType.FUNCTION_LITERAL
+import com.pinterest.ktlint.core.ast.ElementType.LAMBDA_EXPRESSION
 import com.pinterest.ktlint.core.ast.ElementType.LPAR
 import com.pinterest.ktlint.core.ast.ElementType.RPAR
+import com.pinterest.ktlint.core.ast.ElementType.VALUE_ARGUMENT
+import com.pinterest.ktlint.core.ast.ElementType.VALUE_ARGUMENT_LIST
 import com.pinterest.ktlint.core.ast.ElementType.VALUE_PARAMETER
 import com.pinterest.ktlint.core.ast.ElementType.VALUE_PARAMETER_LIST
 import com.pinterest.ktlint.core.ast.ElementType.WHITE_SPACE
@@ -12,6 +16,7 @@ import com.pinterest.ktlint.core.ast.children
 import com.pinterest.ktlint.core.ast.isRoot
 import com.pinterest.ktlint.core.ast.prevLeaf
 import com.pinterest.ktlint.core.ast.visit
+import kotlin.math.max
 import org.jetbrains.kotlin.com.intellij.lang.ASTNode
 import org.jetbrains.kotlin.com.intellij.psi.PsiWhiteSpace
 import org.jetbrains.kotlin.com.intellij.psi.impl.source.tree.LeafElement
@@ -37,7 +42,10 @@ class ParameterListWrappingRule : Rule("parameter-list-wrapping") {
         if (indentSize <= 0) {
             return
         }
-        if (node.elementType == VALUE_PARAMETER_LIST &&
+        val isParameterOrArgumentList = node.elementType == VALUE_PARAMETER_LIST ||
+            // skip if number of arguments is big (we assume it with a magic number of 8)
+            (node.elementType == VALUE_ARGUMENT_LIST && node.children().filter { it.elementType == VALUE_ARGUMENT }.toList().size <= 8)
+        if (isParameterOrArgumentList &&
             // skip when there are no parameters
             node.firstChildNode?.treeNext?.elementType != RPAR &&
             // skip lambda parameters
@@ -48,7 +56,7 @@ class ParameterListWrappingRule : Rule("parameter-list-wrapping") {
             // - maxLineLength exceeded (and separating parameters with \n would actually help)
             // in addition, "(" and ")" must be on separates line if any of the parameters are (otherwise on the same)
             val putParametersOnSeparateLines =
-                node.textContains('\n') ||
+                node.textContainsIgnoringLambda('\n') ||
                     // max_line_length exceeded
                     maxLineLength > -1 && (node.column - 1 + node.textLength) > maxLineLength
             if (putParametersOnSeparateLines) {
@@ -70,10 +78,12 @@ class ParameterListWrappingRule : Rule("parameter-list-wrapping") {
                             }
                         }
                         VALUE_PARAMETER,
+                        VALUE_ARGUMENT,
                         RPAR -> {
                             var paramInnerIndentAdjustment = 0
                             val prevLeaf = child.prevLeaf()
-                            val intendedIndent = if (child.elementType == VALUE_PARAMETER) {
+                            val isParameterOrArgument = child.elementType == VALUE_PARAMETER || child.elementType == VALUE_ARGUMENT
+                            val intendedIndent = if (isParameterOrArgument) {
                                 paramIndent
                             } else {
                                 indent
@@ -108,17 +118,30 @@ class ParameterListWrappingRule : Rule("parameter-list-wrapping") {
                                 }
                             }
                             if (paramInnerIndentAdjustment != 0 &&
-                                child.elementType == VALUE_PARAMETER
+                                isParameterOrArgument
                             ) {
                                 child.visit { n ->
                                     if (n.elementType == WHITE_SPACE && n.textContains('\n')) {
+                                        val isInCollectionOrFunctionLiteral =
+                                            n.treeParent?.elementType == COLLECTION_LITERAL_EXPRESSION || n.treeParent?.elementType == FUNCTION_LITERAL
+
+                                        // If we're inside a collection literal, let's recalculate the adjustment
+                                        // because the items inside the collection should not be subject to the same
+                                        // indentation as the brackets.
+                                        val adjustment = if (isInCollectionOrFunctionLiteral) {
+                                            val expectedPosition = intendedIndent.length + indentSize
+                                            expectedPosition - child.column
+                                        } else {
+                                            paramInnerIndentAdjustment
+                                        }
+
                                         val split = n.text.split("\n")
                                         (n as LeafElement).rawReplaceWithText(
                                             split.joinToString("\n") {
-                                                if (paramInnerIndentAdjustment > 0) {
-                                                    it + " ".repeat(paramInnerIndentAdjustment)
-                                                } else {
-                                                    it.substring(0, Math.max(it.length + paramInnerIndentAdjustment, 0))
+                                                when {
+                                                    it.isEmpty() -> it
+                                                    adjustment > 0 -> it + " ".repeat(adjustment)
+                                                    else -> it.substring(0, max(it.length + adjustment, 0))
                                                 }
                                             }
                                         )
@@ -152,6 +175,8 @@ class ParameterListWrappingRule : Rule("parameter-list-wrapping") {
             LPAR -> """Unnecessary newline before "(""""
             VALUE_PARAMETER ->
                 "Parameter should be on a separate line (unless all parameters can fit a single line)"
+            VALUE_ARGUMENT ->
+                "Argument should be on a separate line (unless all arguments can fit a single line)"
             RPAR -> """Missing newline before ")""""
             else -> throw UnsupportedOperationException()
         }
@@ -165,5 +190,12 @@ class ParameterListWrappingRule : Rule("parameter-list-wrapping") {
             leaf = leaf.prevLeaf()
         }
         return ""
+    }
+
+    private fun ASTNode.textContainsIgnoringLambda(char: Char): Boolean {
+        return children()
+            .flatMap { if (it.elementType == VALUE_ARGUMENT) it.children() else sequenceOf(it) }
+            .filter { it.elementType != LAMBDA_EXPRESSION }
+            .any { it.textContains(char) }
     }
 }